Convert CreditCard::VirtualCardEnrollmentState enum to enum class
Bug: 1418432
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4707109
Commit-Queue: Prakhar Shrivastav <p.shri@samsung.com>
Cr-Commit-Position: refs/heads/main@{#1173416}

+++ components/autofill/core/browser/webdata/autofill_table.cc
@@ -3748,7 +3748,8 @@ void AutofillTable::AddMaskedCreditCards(
     masked_insert.BindInt(index++, static_cast<int>(card.card_issuer()));
     masked_insert.BindString(index++, card.issuer_id());
     masked_insert.BindInt64(index++, card.instrument_id());
-    masked_insert.BindInt(index++, card.virtual_card_enrollment_state());
+    masked_insert.BindInt(
+        index++, static_cast<int>(card.virtual_card_enrollment_state()));
     masked_insert.BindInt(index++, static_cast<int>(
                                    card.virtual_card_enrollment_type()));
     masked_insert.BindString(index++, card.card_art_url().spec());
